Simpler password resets
Take email and new password at the same time and put both in serializer,
meaning user only needs to provide input a single time.

Also means the reset link is tied to a single password and cannot be
used to change to another password than the requested one. It is
possible to request multiple password reset links with different
passwords and use them freely an unlimited amount of times during the
serializers timed validity, resulting in virtually the same behaviour.
Pretty sure that does not result in any actual vulns though, except for
attacks that provide info on the system that we will probably never face
anyway.

Serializer is encrypted with flask secret key, so no need to worry about
sending passwords over insecure channels.