better password reset system, longer token validity
At every password change, a random id is generated. This id is sent to
the user email together with the user id in a timed, signed token. When
the user sends the token to the server, it is first checked for age, and
then if the password id in the token is the same as the one currently in
the users database row. If not, the password has been changed since the
token was generated and the token is therefore invalid. If the password
id is correct, the user is allowed to change the password and a new
password id is generated. The tokens generated with the last password id
are longer valid.

To sum up:
    * Tokens are single-use.
    * Tokens have a time limit.
    * Multiple tokens can be valid at the same time, preventing
        confusion from email delays.
    * Tokens generated before a password change are rendered invalid.
    * Tokens are not stored anywhere, except for in the email.
    * No sensitive information is sent in the email.

Thanks to this question and answer for the idea:
http://crypto.stackexchange.com/a/18262