switch to simplemde and markdown for editing
Markdown is awesome, makes it easy to read and edit the source. Also
lets python handle the rich text output to html instead of a javascript
plugin. The editor is simply a very fancy textbox, even if the js
doesn't work or is disabled, you can write markdown by hand instead.